Refactor ObjectExtensions and CaseInsensitiveComparer
Updated the return value in `ObjectExtensions` to return a
case-insensitive dictionary. Removed the `CaseInsensitiveComparer<TKey>`
class from `HypermediaBase.cs` and reintroduced it in a new file
`CaseInsensitiveComparer.cs` to maintain functionality for
case-insensitive key comparisons.
