+++ IDFCR/IDFCR.Shared.Http/Results/HypermediaBase.cs
@@ -5,19 +5,6 @@ using System.Text.Json.Serialization;
 
 namespace IDFCR.Shared.Http.Results;
 
-public class CaseInsensitiveComparer<TKey> : IEqualityComparer<TKey>
-    where TKey : notnull
-{
-    public bool Equals(TKey? x, TKey? y)
-    {
-        return string.Equals(x?.ToString(), y?.ToString(), StringComparison.OrdinalIgnoreCase);
-    }
-    public int GetHashCode([DisallowNull] TKey obj)
-    {
-        return obj.ToString()?.GetHashCode(StringComparison.OrdinalIgnoreCase) ?? 0;
-    }
-}
-
 public abstract record HypermediaBase : IHypermedia
 {
     private readonly Dictionary<string, object?> _dictionary = new(5, new CaseInsensitiveComparer<string>());
